+++ src/main/java/com/arcbees/bourseje/client/admin/AdminModule.java
@@ -16,21 +16,15 @@
 
 package com.arcbees.bourseje.client.admin;
 
-import com.arcbees.bourseje.client.admin.candidates.CandidatesModule;
-import com.arcbees.bourseje.client.admin.endofvote.EndOfVoteModule;
-import com.arcbees.bourseje.client.admin.howtovote.HowToVoteModule;
-import com.arcbees.bourseje.client.admin.numberofvote.NumberOfVoteModule;
+import com.arcbees.bourseje.client.admin.dashboard.AdminDashboardModule;
 import com.arcbees.bourseje.client.admin.result.ResultModule;
 import com.gwtplatform.mvp.client.gin.AbstractPresenterModule;
 
 public class AdminModule extends AbstractPresenterModule {
     @Override
     protected void configure() {
-        install(new HowToVoteModule());
-        install(new CandidatesModule());
-        install(new NumberOfVoteModule());
+        install(new AdminDashboardModule());
         install(new ResultModule());
-        install(new EndOfVoteModule());
 
         bindPresenter(AdminPresenter.class, AdminPresenter.MyView.class, AdminView.class,
                 AdminPresenter.MyProxy.class);
